plugin_system.mtsを役割ごとに分割 (#2351)
肥大化していたcore/src/plugin_system.mts(3627行)を、内容ごとに以下のファイルへ分割した。
実行時はいずれもplugin_system.mtsにマージされ、単一のプラグイン『plugin_system』として振る舞う。

- plugin_system_debug.mts    : 特殊命令・デバッグ支援・プラグイン管理
- plugin_system_math.mts     : 四則演算・論理演算・ビット演算
- plugin_system_string.mts   : 文字列処理・置換/トリム・文字変換・指定形式・文字種類
- plugin_system_array.mts    : 配列操作・二次元配列処理
- plugin_system_datetime.mts : 日時処理(簡易)
- plugin_system_url.mts      : URLエンコード・パラメータ・BASE64・パス操作

あわせて、snako/cnako3のスタンドアロン実行時に必要なファイルコピー先一覧
(core/command/snako.mts, core/deno/snako.ts, src/cnako3mod.mts)と、
分割ファイルを二重登録しないためのnako_gen.mtsの除外リスト、
命令一覧ドキュメント生成用のbatch/pickup_command.nako3を更新した。

core/test/plugin_system_debug_test.mjs と
core/test/plugin_system_split_test.mjs を追加し、
分割ファイルが正しくplugin_systemへマージされていること、
および代表的な命令が分割後も動作することを確認した。
